fix(storage): address code review — honest verify, i64 sizes, upsert fix
C1: Rename verify endpoint from reconstruction claim to availability check.
Fields `shardsUsedForReconstruction`/`shardsIntentionallySkipped` become
`shardsLocated`/`shardsMissing` across Rust views, Angular service, and HTML.

C2: Widen `total_size_bytes` and `shard_size_bytes` from i32/INTEGER to
i64/BIGINT in migration, diesel schema, models, and http.rs casts so
content larger than 2GB is representable.

I5: Replace `diesel::replace_into` with `insert_or_ignore_into` + conditional
UPDATE in `shard_locations::upsert_location` to preserve `first_seen` timestamps.

+++ elohim/elohim-storage/src/http.rs
@@ -2257,8 +2257,8 @@ impl HttpServer {
                                                     - manifest.data_shards)
                                                     as i32,
                                                 shard_hashes_json: &shard_hashes_json,
-                                                total_size_bytes: manifest.total_size as i32,
-                                                shard_size_bytes: manifest.shard_size as i32,
+                                                total_size_bytes: manifest.total_size as i64,
+                                                shard_size_bytes: manifest.shard_size as i64,
                                                 mime_type: &manifest.mime_type,
                                                 reach: &reach,
                                             };
@@ -2500,8 +2500,8 @@ impl HttpServer {
                                                     - manifest.data_shards)
                                                     as i32,
                                                 shard_hashes_json: &shard_hashes_json,
-                                                total_size_bytes: manifest.total_size as i32,
-                                                shard_size_bytes: manifest.shard_size as i32,
+                                                total_size_bytes: manifest.total_size as i64,
+                                                shard_size_bytes: manifest.shard_size as i64,
                                                 mime_type: &manifest.mime_type,
                                                 reach: &reach,
                                             };
